+++ ports/stm32f4xx/bacnet.h
@@ -1,10 +1,10 @@
-/**************************************************************************
- *
- * Copyright (C) 2010 Steve Karg <skarg@users.sourceforge.net>
- *
- * SPDX-License-Identifier: MIT
- *
- *********************************************************************/
+/**
+ * @file
+ * @brief BACnet stack initialization and task processing
+ * @author Steve Karg
+ * @date 2021
+ * @copyright SPDX-License-Identifier: MIT
+ */
 #ifndef BACNET_H
 #define BACNET_H