+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # BACnet Stack ChangeLog
 
 BACnet open source protocol stack C library for embedded systems,
-Linux, MacOS, BSD, and Windows
+microcontrollers, Linux, MacOS, BSD, and Windows
 
 All notable changes to this project will be documented in this file.
 
@@ -15,9 +15,31 @@ The git repositories are hosted at the following sites:
 ## [Unreleased]
 
 ### Security
+
 ### Added
+
+* Added more API for BACnet basic server device object. (#994)
+* Added the weekly-schedule property write in basic schedule object. (#990)
+* Added uBASIC-Plus program object example to STM32F4xx. (#967)
+* Added guards in create object initialization to prevent memory leaks. (#965)
+
 ### Changed
+
+* Changed mirror script to improve debugging. (#968)
+* Changed dlenv to support multiple datalinks via environment variable. (#966)
+
 ### Fixed
+
+* Fixed NDPU comparison functions that were missing segment-ack PDU. (#991)
+* Fixed WriteProperty NULL bypass which is only for present-value property
+  of commandable objects. (#984)
+* Fixed the ghost Device ID 0 in the I-Am response when the actual
+  routed devices are less than the MAX_NUM_DEVICES for gateway device. (#981)
+* Fixed BACnetGroupChannelValue encoding and decoding of BACnetChannelValue
+  which was deemed errata by BACnet standard committee. (#980)
+* Fixed some INTRINSIC_REPORTING #ifs in AV and BV basic objects. (#977)
+* Fixed network specific original broadcast for IP in apps/router. (#976)(#989)
+
 ### Removed
 
 ## [1.4.1] - 2025-04-11